# flavor-entertainers-backend/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Flavor Entertainers API
  description: Production-ready backend for adult entertainment performer bookings in Western Australia
  version: 1.0.0
  contact:
    name: Flavor Entertainers
    email: contact@lustandlace.com.au
  license:
    name: UNLICENSED
servers:
  - url: https://api.flavor-entertainers.example
    description: Production server
  - url: http://localhost:8080
    description: Development server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "Validation Error"
        message:
          type: string
          example: "Invalid request body"
        code:
          type: string
          example: "VALIDATION_FAILED"
        details:
          type: object
        requestId:
          type: string
          format: uuid
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, performer, client]
        display_name:
          type: string
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Register a new user (client or performer)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - role
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                role:
                  type: string
                  enum: [performer, client]
                display_name:
                  type: string
                phone:
                  type: string
                  pattern: '^\+61[0-9]{9}$'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Registration failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: "bearer"
                  expires_in:
                    type: integer
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/magic-link:
    post:
      tags:
        - Authentication
      summary: Send magic link
      description: Send magic link for passwordless login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Magic link sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...